requirements.txt
Flask==3.1.0
gunicorn==23.0.0
anthropic==0.40.0
google-cloud-storage==2.18.2
python-dotenv==1.0.1
requests==2.32.3
PyMuPDF==1.24.14
PyPDF2==3.0.1